moved Icons to AppConstants
